Rename UserProfileApp
